requirements.txt
torch
transformers
Pillow
pandas
matplotlib
PyYAML
opencv-python-headless
faiss-cpu
git+https://github.com/openai/CLIP.git